An attempt to get around DOMDocument quoting hrefs
The library "CssToInlineStyles" which we use uses DOMDocument under the hood.

DOMDocument does a lot of dumb stuff. One of these dumb things is to "quote" content of href attributes. This seemingly only happens with href attributes, but I don't really know as this seems to be undocumented behaviour.

A test link of `<a href="{{ data.receipt_url}}">` will produce `<a href="%7B%7B%20data.receipt_url%20%7D%7D">`. While `<a hrf="{{ data.receipt_url }}>` will produce `<a hrf="{{ data.receipt_url }}">` as one "might expect".

My workaround is to change all references to href to a sufficiently random string (i banged my keyboard for a while), that is sufficiently unlikely to appear in the original text then feed that text through DOMDocument, then changing the text back to href afterwards.

The objectively better alternative is to remove the plugin all together and instead applying css inlining after the final rendering of the email content. However, it seems a bit more risky as we're not entirely sure what other artifacting DOMDocument could grace us with.